requirements.txt
discord.py>=2.3,<3.0
openai>=1.0,<2.0
python-dotenv>=1.0,<2.0
aiosqlite>=0.19,<1.0
humanize>=4.9,<5.0
aiohttp>=3.9,<4.0
beautifulsoup4>=4.12,<5.0
PyGithub>=2.1,<3.0
fastapi>=0.111,<1.0
uvicorn>=0.29,<1.0
jinja2>=3.1,<4.0
python-multipart>=0.0.9,<1.0
itsdangerous>=2.1,<3.0
qdrant-client>=1.9,<2.0
httpx>=0.27,<1.0
